v1.006xq
5.2凌晨

+++ src/com/xcooper/task/task/web/command/CAjaxUpdateTaskCommand.java
@@ -39,7 +39,7 @@ public class CAjaxUpdateTaskCommand implements Command {
         TaskVO taskVO = taskBean.getTaskByID(id);
 
         //修改 任务名 taskName
-        if(null!=request.getParameter("taskName")&& !("").equals(request.getParameter("taskName"))) {
+        if (null != request.getParameter("taskName") && !("").equals(request.getParameter("taskName"))) {
             taskVO.setTASK_NAME(request.getParameter("taskName"));
         }
         //修改 截止时间 endDatetime
@@ -48,51 +48,54 @@ public class CAjaxUpdateTaskCommand implements Command {
         }
 
         //修改 任务描述 taskInfo
-        if(null!=request.getParameter("taskInfo")) {
+        if (null != request.getParameter("taskInfo")) {
             taskVO.setTASK_INFO(request.getParameter("taskInfo"));
         }
         //修改执行人 exeId
-        if(null!=request.getParameter("exeId")) {
+        if (null != request.getParameter("exeId")) {
             taskVO.setEXE_ID(StrUtil.getNotNullIntValue(request.getParameter("exeId"), 0));
         }
         //修改所属清单 listId
-        if(null!=request.getParameter("listId")) {
+        if (null != request.getParameter("listId")) {
             taskVO.setLIST_ID(StrUtil.getNotNullIntValue(request.getParameter("listId"), 0));
         }
         //修改所属项目 projectId
-        if(null!=request.getParameter("projectId")&& !("").equals(request.getParameter("projectId"))) {
+        if (null != request.getParameter("projectId") && !("").equals(request.getParameter("projectId"))) {
             taskVO.setPROJECT_ID(StrUtil.getNotNullIntValue(request.getParameter("projectId"), 0));
         }
         //修改检查项
         TaskCheckItemBean taskCheckItemBean = new TaskCheckItemBean();
-        TaskCheckItemVO taskCheckItemVO = new TaskCheckItemVO();
 
-        String itemNames = request.getParameter("itemNames");
-        String itemIsDones = request.getParameter("itemIsDones");
+        if (null != request.getParameter("itemNames")) {
+            String itemNames = request.getParameter("itemNames");
+            String itemIsDones = request.getParameter("itemIsDones");
 
-        String[] itemNamesArray = itemNames.split(".,");
-        String[] itemIsDonesArray = itemIsDones.split(".,");
+            String[] itemNamesArray = itemNames.split(".,");
+            String[] itemIsDonesArray = itemIsDones.split(".,");
+            if (!itemNames.equals("")) {
+                MysqlDialect.deleteColl("delete from task_check_item where task_id = " + taskVO.getTASK_ID());
 
-        MysqlDialect.deleteColl("delete from task_check_item where task_id = " + taskVO.getTASK_ID());
+                for (int i = 0; i < itemNamesArray.length; i++) {
 
-        for (int i = 0; i <itemNamesArray.length ; i++) {
+                    TaskCheckItemVO taskCheckItemVO = new TaskCheckItemVO();
 
-            taskCheckItemVO.setID(SeqNumHelper.getNewSeqNum("task_check_item"));
+                    taskCheckItemVO.setID(SeqNumHelper.getNewSeqNum("task_check_item"));
 
-            taskCheckItemVO.setITEM_NAME(itemNamesArray[i]);
+                    taskCheckItemVO.setITEM_NAME(itemNamesArray[i]);
 
-            taskCheckItemVO.setTASK_ID(taskVO.getTASK_ID());
+                    taskCheckItemVO.setTASK_ID(taskVO.getTASK_ID());
 
-            taskCheckItemVO.setIS_DONE(StrUtil.getNotNullIntValue(itemIsDonesArray[i],0));
+                    taskCheckItemVO.setIS_DONE(StrUtil.getNotNullIntValue(itemIsDonesArray[i], 0));
 
-            taskCheckItemBean.addTaskCheckItem(taskCheckItemVO);
+                    taskCheckItemBean.addTaskCheckItem(taskCheckItemVO);
+                }
+            }
         }
-
         //执行修改
         taskBean.modifyTask(taskVO);
 
         //返回ok
-        Collection taskCheckItemList = taskCheckItemBean.getTaskCheckItemColl("select * from task_check_item where task_id = "+ taskVO.getTASK_ID());
+        Collection taskCheckItemList = taskCheckItemBean.getTaskCheckItemColl("select * from task_check_item where task_id = " + taskVO.getTASK_ID());
 
         return JsonResultUtil.instance().addData(taskVO).addExtraData(new Object[]{taskCheckItemList}).json();
     }
